// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  postDates: string[];
  towns: string[];
  prices: string[];
  usersId: string[];
  previewImages: string[];
  offerImages: string[];
  premiumFlags: string[];
  ratings: string[];
  types: string[];
  rooms: string[];
  guests: string[];
  coordinates: string[];
  goods: string[];
  commentsCount: string[];
};